requirements.txt
# local package
# -e .

# external requirements
fastapi==0.95.0
python-dotenv==1.0.0
uvicorn==0.21.1
